pyproject.toml
[tool.poetry]
authors = ["Manuel Lera Ramirez <manulera14@gmail.com>"]
description = ""
name = "all_previous_sgd_peptide_sequences"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.9"
biopython = "^1.79"
requests = "^2.28.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
